update: ログアウトボタンタップ時の処理を実装
